% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\docType{data}
\name{cc_names}
\alias{cc_names}
\title{Valid Firm Names for the Credit Card Data Series}
\format{A character vector of length 13}
\usage{
cc_names
}
\description{
Valid Firm Names for the Credit Card Data Series
}
\keyword{datasets}
