% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{check_api_key}
\alias{check_api_key}
\title{Function to check the validity of an Avenue API key}
\usage{
check_api_key(object)
}
\arguments{
\item{object}{character; typically, this function is called to validate the api_key slot of an object of class AvenueApiClient. It may also be used to examine a length 1 character vector.}
}
\value{
An un-altered api key or, in the case of warnings, message(s) collected from the evaluation process.
}
\description{
Function to check the validity of an Avenue API key
}
