% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{check_date}
\alias{check_date}
\title{Function to check the validity of date inputs}
\usage{
check_date(x, future_warning = FALSE)
}
\arguments{
\item{x}{Date; a length 1 Date or a string expressed in a standard unambiguous format. Descriptive errors are produced for numeric inputs or invalid string formats.}

\item{future_warning}{logical; When \code{TRUE}, a warning message is produced if the supplied date is greater than current date + 1 day.}
}
\value{
When no errors are triggered, check_date returns a valid Date object equal to the input.
}
\description{
Function to check the validity of date inputs
}
\examples{
\dontrun{
check_date(3)
Error in check_date(3) :
'3' is an invalid date selection of class 'numeric'.
You must supply a valid date object or a character of form 'YYYY-MM-DD'!
}
}
