% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{check_dom_source}
\alias{check_dom_source}
\title{Function to check the validity of an Avenue traffic domain}
\usage{
check_dom_source(x)
}
\arguments{
\item{x}{character; a length 1 character vector of the domain you are requesting, e.g. 'google.com'}
}
\value{
When no errors are triggered, check_dom_source returns an appropriate domain name which is passed to the Avenue API.
}
\description{
Function to check the validity of an Avenue traffic domain
}
