% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{check_rev_firm}
\alias{check_rev_firm}
\title{Check that a specificied company name is contained within the 7Parkdata Avenue API for a given revenue data source}
\usage{
check_rev_firm(x, data_source = "", validate_name = NULL)
}
\arguments{
\item{x}{character; a company name}

\item{data_source}{character; a data source name, one of m1, m2, or cc}

\item{validate_name}{logical; when \code{TRUE} (the default) \code{TRUE} is tested against a vector of known/valid names.}
}
\description{
Check that a specificied company name is contained within the 7Parkdata Avenue API for a given revenue data source
}
