% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{check_rev_source}
\alias{check_rev_source}
\title{Function to check the validity of an Avenue Revenue series source input}
\usage{
check_rev_source(x)
}
\arguments{
\item{x}{character; a length 1 character vector and one of "cc" (credit card panel), "m1" (Merchant Intel data), or "m2" (Merchant Intel2 data).}
}
\value{
When no errors are triggered, check_rev_source returns an appropriate dataset name which is passed to the Avenue API.
}
\description{
Function to check the validity of an Avenue Revenue series source input
}
