% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{connect_avenue}
\alias{connect_avenue}
\title{Set credentials for a 7Parkdata Avenue API connection.}
\usage{
connect_avenue(api_key = "")
}
\arguments{
\item{api_key}{character; a valid Avenue API key. Must be a length 1 character vector. Contact your 7Park Sales Representative to obtain your key}
}
\description{
Set credentials for a 7Parkdata Avenue API connection.
}
\examples{
\dontrun{
connection <- connect_avenue(api_key=good_key)
}
}
