% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\docType{data}
\name{extension_names}
\alias{extension_names}
\title{Valid Domain Names for the Avenue Extension Series}
\format{A character vector of length 12,092.}
\usage{
extension_names
}
\description{
Valid Domain Names for the Avenue Extension Series
}
\keyword{datasets}
