% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Helper function to transform an AvenueAPI data object from list to data.frame format}
\usage{
extract_metrics(metrics = metrics, data = NULL)
}
\arguments{
\item{metrics}{character; the metric names of the supplied object}

\item{data}{list; a data set extracted from the .$data slot of an object of class "AvenueAPI'}
}
\value{
a "long" data.frame of the requested metrics
}
\description{
Helper function to transform an AvenueAPI data object from list to data.frame format
}
