% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\docType{methods}
\name{fetch_app_series}
\alias{fetch_app_series}
\alias{fetch_app_series,AvenueApiClient-method}
\alias{fetch_app_series}
\title{fetch_app_series Methods}
\usage{
fetch_app_series(.Object, app, cadence, start_date = "", end_date = "",
  country_code = "", region = NULL)

\S4method{fetch_app_series}{AvenueApiClient}(.Object, app = NULL,
  cadence = NULL, start_date = "", end_date = "", country_code = "",
  region = NULL)
}
\arguments{
\item{.Object}{A valid AvenueApiClient connection}

\item{app}{character; an length-1 vector of the Android package name, e.g. "com.facebook.katana"}

\item{cadence}{character; the cadence requested. Must be one of: 'daily' (daily app usage) or 'weekly' (weekly app usage)}

\item{start_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the start date of the requested series}

\item{end_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the end date of the requested series}

\item{country_code}{character; the ISO 3166 alpha-2 country_code for which you are requesting app data. If left unset, the API returns data for all countries.}

\item{region}{character; the UN region code for which you are requesting app data. An error is thrown if \strong{both} \code{region} and \code{country_code} are set.}
}
\description{
fetch_app_series Methods
}
\examples{
\dontrun{
# Get app usage data for Facebook
fb <- fetch_app_series(gave, app="com.facebook.katana",
                       cadence="daily",
                       start_date='2015-01-01',
                       country_code = 'US')
}
}
