% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\docType{methods}
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_data,AvenueApiClient-method}
\alias{fetch_data}
\title{fetch_data Method.}
\usage{
fetch_data(.Object, ave_url, params = "")

\S4method{fetch_data}{AvenueApiClient}(.Object, ave_url, params = "")
}
\arguments{
\item{.Object}{A valid AvenueApiClient connection}

\item{ave_url}{Fully constructed Avenue API URL}

\item{params}{Params passed to the API}
}
\description{
fetch_data Method.
}
