% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\docType{methods}
\name{fetch_revenue_series}
\alias{fetch_revenue_series}
\alias{fetch_revenue_series,AvenueApiClient-method}
\alias{fetch_revenue_series}
\title{fetch_revenue_series Methods}
\usage{
fetch_revenue_series(.Object, firm, data_source, start_date = "",
  end_date = "", validate_name = TRUE)

\S4method{fetch_revenue_series}{AvenueApiClient}(.Object, firm = NULL,
  data_source = NULL, start_date = "", end_date = "",
  validate_name = TRUE)
}
\arguments{
\item{.Object}{A valid AvenueApiClient connection}

\item{firm}{character; an length-1 vector of the company name, e.g. "Chipotle"}

\item{data_source}{character; the source dataset. Must be one of: 'cc' (credit card panel), 'm1' (Merchant Intel), 'm2' (Merchant Intel2)}

\item{start_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the start date of series}

\item{end_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the end date of series}

\item{validate_name}{logical; when \code{TRUE}, AvenueAPI checks the supplied firm name against a vector of known/valid names.}
}
\description{
fetch_revenue_series Methods
}
\examples{
\dontrun{
# Get credit card data for Chipotle
chip <- fetch_revenue_series(connection,
        firm="Chipotle",
        data_source="cc",
        start_date='2015-01-01')
}
}
