% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\docType{methods}
\name{fetch_traffic_series}
\alias{fetch_traffic_series}
\alias{fetch_traffic_series,AvenueApiClient-method}
\alias{fetch_traffic_series}
\title{fetch_traffic_series Methods}
\usage{
fetch_traffic_series(.Object, domain = NULL, platform = "",
  dataseries = "", start_date = "", end_date = "", country_code = "",
  validate_name = TRUE)

\S4method{fetch_traffic_series}{AvenueApiClient}(.Object, domain = NULL,
  platform = "PC", dataseries = "extension", start_date = "",
  end_date = "", country_code = "US", validate_name = TRUE)
}
\arguments{
\item{.Object}{A valid AvenueApiClient connection}

\item{domain}{character; an length-1 vector of the requested domain name, e.g. "www.google.com"}

\item{platform}{character; the user's computing platform. Must be one of: 'PC' (desktop browser data), 'mobile' (mobile web browser data), or 'ALL' (mobile + PC).}

\item{dataseries}{character; currently, the only valid value for this parameter is "extension" (the default). This will be updated as new traffic series are added to the API.}

\item{start_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the start date of the requested series}

\item{end_date}{string; a length-one character vector expressed as '\code{YYYY-MM-DD}' (ISO 8601) representing the end date of the requested series}

\item{country_code}{character; the ISO 3166 alpha-2 country_code for which you are requesting traffic data. If left unset, the API returns data for all countries.}

\item{validate_name}{logical; when \code{TRUE}, AvenueAPI checks the supplied domain name against a vector of known/valid names.}
}
\description{
fetch_traffic_series Methods
}
\examples{
\dontrun{
# Get domain traffic data for google.com, US
goog   <- fetch_traffic_series(conection, domain = 'google.com',
                               platform = 'PC',
                               start_date='2014-01-01',
                               country_code = 'US',
                               validate_name = TRUE)
}
}
