% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\docType{data}
\name{m1_names}
\alias{m1_names}
\title{Valid Firm Names for the Merchant Intel Data Series}
\format{A character vector of length 1,311}
\usage{
m1_names
}
\description{
Valid Firm Names for the Merchant Intel Data Series
}
\keyword{datasets}
