% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_resources.R
\docType{data}
\name{m2_names}
\alias{m2_names}
\title{Valid Firm Names for the Merchant Intel2 Data Series}
\format{A character vector of length 571}
\usage{
m2_names
}
\description{
Valid Firm Names for the Merchant Intel2 Data Series
}
\keyword{datasets}
