% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avenue_Auth_Methods.R
\name{transform_avenue_series}
\alias{transform_avenue_series}
\title{Transform a 7Parkdata Avenue API return into a "long" or "wide" dataframe for anaysis.}
\usage{
transform_avenue_series(x, wide = FALSE)
}
\arguments{
\item{x}{list; a 7Park Data revenue object pulled with \code{fetch_*_series} method.}

\item{wide}{logical; if \code{TRUE}, a "wide" data frame is returned using \code{\link{dcast}}}
}
\description{
Transform a 7Parkdata Avenue API return into a "long" or "wide" dataframe for anaysis.
}
